Get rid of doctree::Function
